requirements.txt
docopt
pytaglib